mixins: usb: Enable autosuspend for USB host and USB gadget
Currently the XHCI and XDCI ports do not get runtime suspended
even when we do not have any device/host connected to them. This
is because the power/control attribute for them is by default
being set to 'on'. For autosuspend to occur, the control attribute
needs to be 'auto'. Modify the sysfs power/control aattribute so
that these ports may get suspended when not in use and thus not
act as a blocker for S0ix.

JIRA: AIA-242
Test: Boot to home screen. Disconnect Type-C cable being used for
adb access. Disconnect any hub/keyboard/mouse connected to xHCI
port. Check the telemetry stats for XDCI/XHCI gating status. Also,
check the runtime_suspended_time for for pci device 15.0 and 15.1.

Signed-off-by: Abhilash Kesavan <abhilash.kesavan@intel.com>