Boot Optimization: Custom Bootanimation zip file
The default bootanimation.zip file contained images & sequence in which they
are displayed to animate till bootcomplete. But it was seen that even after
bootcomplete there were 98 to 157 frames left in animation. At 30fps, this
meant a delay of 3.2 to 5.2sec to display the home screen even after background
threads are completed. The 2sec uncertainty in delay is due to the 60frame
animation loop which runs till bootcomplete. For combating this there were 2
options, either to completely remove animation for achieving best boottime, or
to keep considerable animation without much CPU overhead & ZERO post
bootcomplete delay. As Celadon is a development platform, and developers
expect something on the bootscreen to infer booting, we applied the later
solution, to get close to best results keeping minimal animation.

With this bootanimation, following changes will appear:
- rotating balls loop will not appear
- formation of android logo from 4balls will take place pre-bootcomplete
- transition of android logo to black screen removed
- after animation is over, android logo will remain on the screen till
bootcomplete

Following were the scenarios that were tested along with data on APL-NUC:
+---+---------------------------------------------------+------------------+
| # |           Description                             |    Boot-time     |
| 1 | No Boot Animation (Black screen)                  | 18.94 s          |
| 2 | Single Frame (ANDROID logo)                       | 19.30 s          |
| 3 | Above mentioned Animation w/ jerky home transition| 19.36 s <-Applied|
| 4 | Above Animation with smooth home screen transition| 20.62 s          |
| 5 | Default Animation w/o any change                  | 25.45 s          |
+---+---------------------------------------------------+------------------+

The default zip file is present in system/media, the custom bootanimation.zip
file is being copied to product/media using this patch because product/media
directory takes precedence over system/media.

Tracked-On: OAM-72177
Signed-off-by: Vishal Sharma <vishal1.sharma@intel.com>