Add support for adding directories in mixin groups and build flash files
[CELADON] Add support for adding directories in mixin groups and build flash files

Currently, for mixins, only individual files are picked up by the build system.
There is no support for adding one or more files inside a sub-directory for any mixin group.

A change in the build system is required to incorporate addition of directories as part of mixins and also as part of the flash files generated after build completion.

Changes have been made for creation of flashfiles zip if directories are encountered instead of files

In the current file, recursive copy of files inside a directory for creation of intermediate flashfiles zip has been enabled

Tracked-On: OAM-84419
Signed-off: debarghy <debarghya.bhattacharya@intel.com>