Upgrade target-level to 4 in Android 10
This patch fixes the target level required for
Android 10 to 4 in device manifest.

Changing the target-level to 4 will require the below
changes:
- The kernel config is the dependency for the target
  level 4.
- graphics mapper HAL version needed for target level 4
  is >= 2.1

Shipping FCM Version should be >= 4 if Shipping API level == 29.
This is the hard requirement for Android 10 set by Google.
This can be found in the Google GMS requirement document.

Tracked-On: OAM-92918
Signed-off-by: Tanuj Tekriwal <tanuj.tekriwal@intel.com>