Set properties to use new volume metadata encryption method
In AOSP, there are two implementations of metadata encryption
on adoptable storage: a deprecated one based on dm-crypt and
a newer one based on dm-default-key. To ensure that the correct
implementation is selected, either api level needs to be set to
30 or new volume metadata encryption method, format and options
properties needs to be set.

AdoptableHostTest CTS tests are failing as neither api level nor
properties set to use new metadata encryption method.

Fix the issue by setting properties to use new metadata
encryption method.

Tracked-On: OAM-95556
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>