[WA]audio server restart on boot complete
Since hda driver is unable to get the HDMI
sound devices, it is waiting till the timeout
and then listing the ALC256 cards.
By then Android would have started the Audioserver
and finds that there are no sink devices availabe.
Till we get the HDMI solution working this can be
a WA fix for the audio 3.5mm playback

Tracked-On: OAM-95721
Signed-off-by: gkdeepa g.k.deepa@intel.com