Move persist.adb.tcp.port from Vendor to Product Partition
As per GMS requirements document,for devices launching with
Android 11 or higher, the names of property contexts of vendor
(or odm) properties MUST start with vendor_ or odm_.

Keeping persist.adb.tcp.port in vendor partition and exporting it
to system partition using exported2_system_prop will violate this
requirement and cause VTS failures. Since this property is used
in system partition but is owned by vendor the correct way is
adding it in Product partition.

This patch also adds logic to mount  EP0 only when adb over
usb is supported, otherwise adb devices will not work in GSI Image.

Tracked-On: OAM-95781
Signed-off-by: pmandri <padmasree.mandri@intel.com>
Signed-off-by: svenate <salini.venate@intel.com>
Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>