Disable signing of modules as its not mandatory for Android
Loadable kernel modules are not getting loaded with
boot-debug.img flashed. Loading of module is failing with
"Loading of module with unavailable key is rejected" error.

Following is the doc snippet from Android loadable kernel
modules->Module signing section

On devices required to support verified boot, Android requires
kernel modules to be in the partitions that have dm-verity
enabled. Module signing isn't mandatory and isn't tested against;
however, if desired, an ODM can enable module signing as long as
they have the key signing and other infrastructure required to
ensure independent kernel and file system OTA updates in the
future.

As module signing is not mandatory, disabling the signing of
modules to make sure loadable  modules are propertly getting
loaded with boot-debug.img.

Tracked-On: OAM-96153
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>