Plugin to regulate the host suspend
Inhibitor lock is used to block the host suspend requests by user.
This will stop host suspend when the Android screen is on

This plugin inhibits the suspend operation of the host
based on the guest screen ON/OFF status.

Plugin will be downloaded to scripts folder of caas during
the host setup.

Tracked-On: OAM-97461
Signed-off-by: Shwetha B <shwetha.b@intel.com>