Update kernel configs for hda and sof selection using android property
Audio playback and audio recording not working when sof is enabled.

Inorder to support sof and hda selection using android property,
some of the drivers had to be loaded dynamically. Since some of the
drivers are static, switching between sof and hda using android
property is not working.

Fix the issue by loading the needed sound specific drivers dynamically
instead of static.

Tracked-On: OAM-104466
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>