Fix for wifi not working due to switch from hidl to aidl
Wi-Fi not working as the hwservicemanager is not able to find
android.hardware.wifi@1.0::IWifi/default.

hwservicemanager is not able to find the service as the Wi-Fi service
is switched to aidl and service name has changed too.

Changes done to fix the issue:
- Change the service name to aidl service
- Remove the wifi related VINTF entries from framework_manifest
as those are included as part of respective vendor service binaries.

Tracked-On: OAM-113175
Signed-off-by: Badrappan, Jeevaka <jeevaka.badrappan@intel.com>