linux-intel-lts2021: enable dummy cpufreq driver
Enabling dummy cpufreq driver for guest vm kernel.

cpufreq drivers are responsible for managing the frequency scaling
of a CPU, allowing it to dynamically adjust its clock speed based on
the system's workload. This feature is essential for optimizing power
consumption and performance on modern processors.

When it comes to VMs, there are a few possible reasons why cpufreq
drivers might not be loaded due to Hypervisor Limitations. The VM's
hypervisor, which is the software responsible for managing the VM,
might not provide full access to all CPU features, including frequency
scaling capabilities.

When a specific cpufreq governor, such as Intel P-state or ACPI cpufreq
(acpi or pcc), is not available, then it is recommended to use
CONFIG_CPUFREQ_DUMMY for dummy cpufreq driver as a fallback mechanism.
It ensures that even if no real frequency scaling driver is
present, the cpufreq framework can still operate without throwing errors
or breaking the system or applications.

Tracked-On: OAM-111324
Signed-off-by: raju <raju.mallikarjun.chegaraddi@intel.com