Add inlinecrypt mount option for data partition
ext4 won't adopt direct_io if one file is needed to be encrypted
and the inlinecrypt mount flag is not declared, it will fallback
to buffer io if the O_DIRECT flag is set by userspace. This will
result in the abnormal memory test scores for Antutu on RPL IVI
Board, since Antutu's random read/write test need the direct_io
support.

Tracked-On: OAM-112014
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>