Use bootloader image from target files for flashfiles package
Currently the bootloader image of flashfiles is generated during
the flashfiles creation, and it has a little difference with the
bootloader image from target files. The bootloader image of
flashfiles is originally created for the UEFI and contain some
additional files which are not used by the SBL, the bootloader
image from target files is used for the SBL and remove the files
which are not used by the SBL. The incremental ota pacakge is
generated from the target files, and this make the incremental
ota doesn't work on current bootloader image used by the device.

Error message during the incremental ota:
The hash of the source data on disk for this operation doesn't
match the expected value. This could mean that the delta update
payload was targeted for another version, or that the source
partition was modified after it was installed.

Test done: Incremental ota works as expectation.

Tracked-On: OAM-113074
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>