Fix for Null Subscription Manager
The Null Subscription Manager issue arises due to the absence of
the required telephony subscription features in the build
configuration for automotive devices.

Introduced a new XML file, `android.hardware.telephony.automotive.xml`,
which defines the standard set of features required for devices to
support the Telephony Subscription API. This ensures that the telephony
features are included in the product build.

Tests done:
1. Verified that the new XML file is correctly copied to the
`vendor/etc/permissions` directory during the build process.
2. Ensured that the telephony subscription features are recognized
and functional on the target automotive device.
3. Sms are synced from the reference Pixel Device after the BT connection
is successful.

Tracked-On: OAM-128577
Signed-off-by: Bhadouria, Aman <aman.bhadouria@intel.com>
Co-authored-by: Panda, Bharat B <bharat.b.panda@intel.com>
Co-authored-by: Babu, Gowtham Anandha <gowtham.anandha.babu@intel.com>