Remove logwrapper for auto_detection
logwrapper is a utility in Android that wraps a given command and
redirects its output (both stdout and stderr) to the Android
logging system. It's usually used in early development stage,
and for auto_detection, logwrapper is here for serveral years
now. From the SELinux's perspective, it's very confusing why a
log utility needs some much permissions.

Tracked-On: OAM-128889
Signed-off-by: Zhenlong Z Ji <zhenlong.z.ji@intel.com>